requirements.txt
boto==2.48.0
boto3==1.7.31
botocore==1.10.31
bz2file==0.98
certifi==2018.4.16
chardet==3.0.4
cycler==0.10.0
docutils==0.14
gensim==3.4.0
glove-python==0.1.0
idna==2.6
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==2.2.2
nltk==3.3
numpy==1.14.3
pandas==0.23.0
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
smart-open==1.5.7
urllib3==1.22
